+++ vllm/entrypoints/openai/serving_embedding.py
@@ -40,7 +40,7 @@ from vllm.outputs import (
 )
 from vllm.pooling_params import PoolingParams
 from vllm.utils import chunk_list
-from vllm.utils.async_utils import merge_async_iterators
+from vllm.utils.asyncio import merge_async_iterators
 
 logger = init_logger(__name__)
