+++ vllm/executor/ray_distributed_executor.py
@@ -24,7 +24,7 @@ from vllm.utils import (
     get_ip,
     get_open_port,
 )
-from vllm.utils.async_utils import make_async
+from vllm.utils.asyncio import make_async
 from vllm.v1.outputs import SamplerOutput
 
 if ray is not None:
