+++ vllm/model_executor/models/interfaces.py
@@ -24,7 +24,7 @@ from vllm.inputs import TokensPrompt
 from vllm.inputs.data import PromptType
 from vllm.logger import init_logger
 from vllm.model_executor.layers.quantization import QuantizationConfig
-from vllm.utils.func import supports_kw
+from vllm.utils.functools import supports_kw
 
 from .interfaces_base import VllmModel, is_pooling_model
