+++ vllm/multimodal/processing.py
@@ -26,7 +26,7 @@ from vllm.logger import init_logger
 from vllm.transformers_utils.processor import cached_processor_from_config
 from vllm.transformers_utils.tokenizer import AnyTokenizer, decode_tokens, encode_tokens
 from vllm.utils import flatten_2d_lists, full_groupby
-from vllm.utils.func import get_allowed_kwarg_only_overrides
+from vllm.utils.functools import get_allowed_kwarg_only_overrides
 from vllm.utils.jsontree import JSONTree, json_map_leaves
 
 from .hasher import MultiModalHasher
