+++ vllm/utils/functools.py
